Changed data type in instruction struct

+++ src/chip8/instructions.rs
@@ -3,17 +3,17 @@ pub struct Instruction {
     /// Raw unparsed instruction
     pub raw_instr: u16,
     /// Highest 4 bits
-    pub op: u16,
+    pub op: u8,
     /// Lowest 12 bits
     pub nnn: u16,
     /// Lowest 4 bits
-    pub n: u16,
+    pub n: u8,
     /// Lower 4 bits of high byte of instruction
-    pub x: u16,
+    pub x: u8,
     /// Upper 4 bits of low byte of instruction
-    pub y: u16,
+    pub y: u8,
     /// Lowest 8 bits of instruction
-    pub kk: u16,
+    pub kk: u8,
 }
 
 /// CHIP-8 Instruction Set
